Move formula validation tests into a separate file
